CMS-2348 Index-operations for all node-operations

+++ modules/wem-core/src/main/java/com/enonic/wem/core/entity/DeleteNodeByPathHandler.java
@@ -1,16 +1,28 @@
 package com.enonic.wem.core.entity;
 
+import javax.inject.Inject;
 import javax.jcr.Session;
 
 import com.enonic.wem.api.command.entity.DeleteNodeByPath;
 import com.enonic.wem.api.command.entity.DeleteNodeResult;
 import com.enonic.wem.api.entity.NoNodeAtPathFound;
+import com.enonic.wem.api.entity.Node;
 import com.enonic.wem.core.command.CommandHandler;
 import com.enonic.wem.core.entity.dao.NodeJcrDao;
+import com.enonic.wem.core.index.IndexService;
 
 public class DeleteNodeByPathHandler
     extends CommandHandler<DeleteNodeByPath>
 {
+
+    private IndexService indexService;
+
+    @Inject
+    public void setIndexService( final IndexService indexService )
+    {
+        this.indexService = indexService;
+    }
+
     @Override
     public void handle()
         throws Exception
@@ -18,13 +30,22 @@ public class DeleteNodeByPathHandler
         final Session session = context.getJcrSession();
         final NodeJcrDao itemDao = new NodeJcrDao( session );
 
+        final Node nodeByPath = itemDao.getNodeByPath( command.getPath() );
+
+        // TODO: How should we handle delete by path in index? For now fetch node and get Id here
+        if ( nodeByPath == null )
+        {
+            command.setResult( DeleteNodeResult.NOT_FOUND );
+            return;
+        }
+
         try
         {
             itemDao.deleteNodeByPath( command.getPath() );
             session.save();
             command.setResult( DeleteNodeResult.SUCCESS );
 
-            // TODO: delete from index
+            indexService.deleteEntity( nodeByPath.id() );
         }
         catch ( NoNodeAtPathFound e )
         {
